Implement time picker for make target timed drill as well
