+++ lib/screens/active/widgets/timed_drill_widget.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import '../../../models/drill.dart';
 import '../../../services/workout_state.dart';
 import 'package:flutter_bball_app/utils/format_duration.dart';
-import 'package:numberpicker/numberpicker.dart';
+import 'package:flutter_bball_app/screens/active/widgets/time_picker_dialog.dart' as custom_picker;
 
 class TimedDrillWidget extends StatefulWidget {
   final Drill drill;
@@ -104,7 +104,7 @@ class _TimedDrillWidgetState extends State<TimedDrillWidget> {
                 }
                 final result = await showDialog<int>(
                   context: context,
-                  builder: (context) => _TimePickerDialog(initialSeconds: _remainingSeconds),
+                  builder: (context) => custom_picker.TimePickerDialog(initialSeconds: _remainingSeconds),
                 );
                 if (result != null) {
                   setState(() {
@@ -137,73 +137,6 @@ class _TimedDrillWidgetState extends State<TimedDrillWidget> {
             color: Color(0xFF9ca3af),
           ),
           textAlign: TextAlign.center,
-        ),
-      ],
-    );
-  }
-}
-
-class _TimePickerDialog extends StatefulWidget {
-  final int initialSeconds;
-  const _TimePickerDialog({Key? key, required this.initialSeconds}) : super(key: key);
-
-  @override
-  State<_TimePickerDialog> createState() => _TimePickerDialogState();
-}
-
-class _TimePickerDialogState extends State<_TimePickerDialog> {
-  int minutes = 0;
-  int seconds = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    minutes = widget.initialSeconds ~/ 60;
-    seconds = (widget.initialSeconds % 60) - (widget.initialSeconds % 5);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      backgroundColor: const Color(0xFF1f2937),
-      title: const Text('Set Timer', style: TextStyle(color: Colors.white)),
-      content: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          NumberPicker(
-            value: minutes,
-            minValue: 0,
-            maxValue: 59,
-            zeroPad: true,
-            itemHeight: 60,
-            onChanged: (value) => setState(() => minutes = value),
-            textStyle: const TextStyle(color: Colors.white54, fontSize: 32),
-            selectedTextStyle: const TextStyle(color: Colors.white, fontSize: 48, fontWeight: FontWeight.bold),
-          ),
-          const Text(':', style: TextStyle(color: Colors.white, fontSize: 48)),
-          NumberPicker(
-            value: seconds,
-            minValue: 0,
-            maxValue: 55,
-            step: 5,
-            zeroPad: true,
-            itemHeight: 60,
-            onChanged: (value) => setState(() => seconds = value),
-            textStyle: const TextStyle(color: Colors.white54, fontSize: 32),
-            selectedTextStyle: const TextStyle(color: Colors.white, fontSize: 48, fontWeight: FontWeight.bold),
-          ),
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.of(context).pop(),
-          child: const Text('Cancel', style: TextStyle(color: Color(0xFF9ca3af))),
-        ),
-        TextButton(
-          onPressed: () {
-            Navigator.of(context).pop(minutes * 60 + seconds);
-          },
-          child: const Text('Set', style: TextStyle(color: Color(0xFF3b82f6), fontWeight: FontWeight.bold)),
         ),
       ],
     );
